traffic/smart-city-sustainability-platform/backend/requirements.txt
# FastAPI関連
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# データベース関連
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# InfluxDB関連
influxdb-client==1.38.0

# Kafka関連
kafka-python==2.0.2
confluent-kafka==2.3.0

# 認証・セキュリティ
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTPクライアント
httpx==0.25.2
aiohttp==3.9.1

# データ分析・可視化
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0

# 機械学習（オプション）
# tensorflow==2.15.0
# torch==2.1.1

# レポート生成
reportlab==4.0.7
weasyprint==60.2
openpyxl==3.1.2
jinja2==3.1.2

# 日時処理
python-dateutil==2.8.2
pytz==2023.3

# ログ
loguru==0.7.2

# テスト
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# その他
python-dotenv==1.0.0
pydantic-email-validator==2.0.0

